Eglibc and perhaps other libcs are dependent on dynamic loading of external sublibraries (NSS) for gethostbyname() to properly function, and they are lazily loaded (via dlopen()) on the first call by a process.  This is problematic when running from a chroot, where these libraries will not be available if they have not been manually added to the directory structure.
However, once these libraries are loaded, they are now part of the address
space of the process and will continue to be accessible, thanks to the magic of
hardware virtual memory and the kernel.  This fact provides a simple
workaround.  Just make a worthless invocation to gethostbyname() before
chrooting in order to force this process to happen while it can succeed.